overlay/: fix password-protected hotspot on 24.04-1.x
Set NetworkManager's `wifi-wpa-psk-pmf-no-optional-ap=yes` workaround
to fix password-protected hotspot failing to be connected while running
24.04-1.x.

I'm not sure why 20.04 doesn't need this; maybe some change in
wpa_supplicant in Ubuntu 24.04?

Now, I would like to express a disappointment to myself that I don't
find a time to implement NetworkManager upstream's suggestion on a more
appropriate config knob [^1]. This is one of the reason I tend to push
back "temporary fix": once it's in place, the incentive to actually fix
it properly is very low, and "temporary fix" becomes very permanent as
time goes on.

[^1]: https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/issues/1324#note_1987861

Fixes: https://gitlab.com/ubports/porting/reference-device-ports/android11/fairphone-4/fairphone-fp4/-/issues/89